calcPython/requirements.txt
hypothesis==6.100.0
pytest
PyYAML==6.0.3